+++ packages/components/pages/f-checkout/test-utils/component-objects/f-checkout-error.component.js
@@ -1,0 +1,14 @@
+import Page from '@justeat/f-wdio-utils';
+import {
+    ERROR_PAGE_COMPONENT
+} from './f-checkout-selectors';
+
+class CheckoutError extends Page {
+    constructor () {
+        super('page', 'checkout-component');
+    }
+
+    get component () { return $(ERROR_PAGE_COMPONENT); }
+}
+
+export default new CheckoutError();