+++ packages/components/pages/f-checkout/test/component/f-checkout.component.spec.js
@@ -1,8 +1,6 @@
-import forEach from 'mocha-each';
+import Checkout from '../../test-utils/component-objects/f-checkout.component';
 
-const Checkout = require('../../test-utils/component-objects/f-checkout.component');
-
-let checkout;
+let tests;
 
 const checkoutInfo = {
     serviceType: 'delivery',
@@ -14,103 +12,97 @@ const checkoutInfo = {
 const orderTime = 'Wednesday 01:45';
 
 describe('f-checkout component tests', () => {
-    beforeEach(() => {
-        // Arrange
-        checkout = new Checkout();
-
+    beforeEach(async () => {
         // Act
-        checkout.load({
+        await Checkout.load({
             ...checkoutInfo
         });
     });
 
-    it('should submit the checkout form', async () => {
+    it.skip('should submit the checkout form', async () => {
         // Arrange
         const orderNote = 'Doorbell is broken';
         // Act
-        await checkout.setFieldValue('orderNote', orderNote);
-        await checkout.selectOrderTime(orderTime);
-        await checkout.goToPayment();
+        await Checkout.setFieldValue('orderNote', orderNote);
+        await Checkout.selectOrderTime(orderTime);
+        await Checkout.goToPayment();
 
         // Assert
         // Waiting for route here, so we can grab redirect url and show form submits.
     });
 
-    forEach([
-        [100, 'addressLine1'],
-        [100, 'addressLine2'],
-        [50, 'addressLocality'],
-        [50, 'addressPostcode'],
-        [16, 'mobileNumber']
-    ])
-        .it('should prevent a user from entering more than "%s" characters in the "%s" field', async (maxlength, field) => {
+    tests = [
+        { maxLength: 100, field: 'addressLine1' },
+        { maxLength: 100, field: 'addressLine2' },
+        { maxLength: 50, field: 'addressLocality' },
+        { maxLength: 50, field: 'addressPostcode' },
+        { maxLength: 16, field: 'mobileNumber' }
+    ];
+
+    tests.forEach(({ maxLength, field }) => {
+        it(`should prevent a user from entering more than "${maxLength}" characters in the "${field}" field`, async () => {
             // Arrange
-            await checkout.clearBlurField(field);
-            const userEntry = 'A'.repeat(maxlength + 1); // Enter more than allowed
+            await Checkout.clearBlurField(field);
+            const userEntry = 'A'.repeat(maxLength + 1); // Enter more than allowed
 
             // Act
-            await checkout.setFieldValue(field, userEntry);
-            await checkout.goToPayment();
+            await Checkout.setFieldValue(field, userEntry);
+            await Checkout.goToPayment();
 
             // Assert
-            await expect((await checkout.getFieldValue(field)).length).toEqual(maxlength);
+            await expect((await Checkout.getFieldValue(field)).length).toEqual(maxLength);
         });
+    });
 
     it.skip('should enable a user to submit without adding a note', async () => {
         // Act
-        await checkout.selectOrderTime(orderTime);
-        await checkout.goToPayment();
+        await Checkout.selectOrderTime(orderTime);
+        await Checkout.goToPayment();
 
         // Assert
         // Waiting for route here, so we can grab redirect url and show form submits.
     });
 
     it('should close the checkout error when "Retry" is clicked', async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             patchCheckoutError: 'restaurant-not-taking-orders'
         });
 
-        await checkout.goToPayment();
-        await checkout.clickRetryButton();
+        await Checkout.goToPayment();
+        await Checkout.clickRetryButton();
         await browser.pause(2000);
 
         // Assert
-        await expect(await checkout.isCheckoutErrorMessageDisplayed()).toBe(false);
+        await expect(await Checkout.isCheckoutErrorMessageDisplayed()).toBe(false);
     });
 
     describe('when the "Duplicate Order Warning" modal is displayed', () => {
         beforeEach(async () => {
-            // Arrange
-            checkout = new Checkout();
-
             // Act
-            checkout.load({
+            await Checkout.load({
                 ...checkoutInfo,
                 placeOrderError: 'duplicate'
             });
 
-            await checkout.goToPayment();
+            await Checkout.goToPayment();
         });
 
         it('should close the modal and remain on the "Checkout Page" when the "Close" button is pressed', async () => {
             // Act
-            await checkout.waitForComponent();
-            await checkout.clickRetryButton();
+            await Checkout.waitForComponent();
+            await Checkout.clickRetryButton();
 
             // Assert
-            await expect(await checkout.isCheckoutErrorMessageDisplayed()).toBe(false);
-            await expect(await checkout.isComponentDisplayed()).toBe(true);
+            await expect(await Checkout.isCheckoutErrorMessageDisplayed()).toBe(false);
+            await expect(await Checkout.isComponentDisplayed()).toBe(true);
         });
 
         it('should attempt to redirect to the "Order History Page" when the "View my orders" button is pressed', async () => {
             // Act
-            await checkout.waitForComponent();
-            await checkout.clickDupOrderGoToHistoryButton();
+            await Checkout.waitForComponent();
+            await Checkout.clickDupOrderGoToHistoryButton();
 
             // Assert
             await expect((await browser.getUrl()).split('/')[3]).toBe('order-history');