+++ packages/components/pages/f-contact-preferences/test/accessibility/axe-accessibility.spec.js
@@ -1,54 +1,45 @@
-import forEach from 'mocha-each';
-import { getAxeResults } from '../../../../../../test/utils/axe-helper'; // eslint-disable-line import/no-relative-packages
-
-const ContactPreferences = require('../../test-utils/component-objects/f-contactPreferences.component');
-
-let contactPreferences;
+import ContactPreferences from '../../test-utils/component-objects/f-contactPreferences.component';
+import ContactPreferencesError from '../../test-utils/component-objects/f-contactPreferences-error.component';
 
 describe('Accessibility tests', () => {
-    beforeEach(() => {
-        // Arrange
-        contactPreferences = new ContactPreferences();
-    });
+    const tests = [
+        'en-GB'
+    ];
 
-    forEach([
-        ['en-GB']
-    ]).it('a11y - should test that the %s f-contact-preferences component is WCAG compliant', locale => {
-        // Act
-        contactPreferences.load({ locale });
-        const axeResults = getAxeResults('f-contact-preferences');
+    tests.forEach(locale => {
+        it(`a11y - should test that the ${locale} f-contact-preferences component is WCAG compliant`, async () => {
+            // Act
+            await ContactPreferences.load({ locale });
+            const axeResults = await ContactPreferences.getAxeResults('f-contact-preferences');
 
-        // Assert
-        expect(axeResults.violations.length).toBe(0);
-    });
-
-    forEach([
-        ['en-GB']
-    ]).it('should test that the %s f-contact-preferences (error alert) is WCAG compliant', locale => {
-        // Act
-        contactPreferences.load({
-            locale,
-            apiState: 'post-preferences-fails'
+            // Assert
+            await expect(axeResults.violations.length).toBe(0);
         });
-        contactPreferences.clickNewsEmailCheckbox();
-        contactPreferences.clickSubmitButton();
-        const axeResults = getAxeResults('f-contact-preferences - save error alert');
 
-        // Assert
-        expect(axeResults.violations.length).toBe(0);
-    });
+        it(`should test that the ${locale} f-contact-preferences (error alert) is WCAG compliant`, async () => {
+            // Act
+            await ContactPreferences.load({
+                locale,
+                apiState: 'post-preferences-fails'
+            });
+            await ContactPreferences.clickNewsEmailCheckbox();
+            await ContactPreferences.clickSubmitButton();
+            const axeResults = await ContactPreferences.getAxeResults('f-contact-preferences - save error alert');
 
-    forEach([
-        ['en-GB']
-    ]).it('should test that the %s f-contact-preferences (error page) is WCAG compliant', locale => {
-        // Act
-        contactPreferences.load({
-            locale,
-            apiState: 'get-preferences-fails'
+            // Assert
+            await expect(axeResults.violations.length).toBe(0);
         });
-        const axeResults = getAxeResults('f-contact-preferences - load error page');
 
-        // Assert
-        expect(axeResults.violations.length).toBe(0);
+        it('should test that the %s f-contact-preferences (error page) is WCAG compliant', async () => {
+            // Act
+            await ContactPreferencesError.load({
+                locale,
+                apiState: 'get-preferences-fails'
+            });
+            const axeResults = await ContactPreferencesError.getAxeResults('f-contact-preferences - load error page');
+
+            // Assert
+            await expect(axeResults.violations.length).toBe(0);
+        });
     });
 });