+++ packages/components/pages/f-registration/test-utils/component-objects/f-registration.selectors.js
@@ -1,17 +1,33 @@
-exports.CREATE_ACCOUNT_BUTTON = '[data-test-id="create-account-submit-button"]';
-exports.TERMS_AND_CONDITIONS_LINK = '[data-test-id="ts-and-cs-link"]';
-exports.PRIVACY_POLICY_LINK = '[data-test-id="privacy-policy-link"]';
-exports.COOKIES_POLICY_LINK = '[data-test-id="cookies-policy-link"]';
-exports.REQUEST_ERROR = '[data-test-id="error-message-component"]';
+const CREATE_ACCOUNT_BUTTON = '[data-test-id="create-account-submit-button"]';
+const TERMS_AND_CONDITIONS_LINK = '[data-test-id="ts-and-cs-link"]';
+const PRIVACY_POLICY_LINK = '[data-test-id="privacy-policy-link"]';
+const COOKIES_POLICY_LINK = '[data-test-id="cookies-policy-link"]';
+const REQUEST_ERROR = '[data-test-id="error-message-component"]';
 
-exports.FIRST_NAME_INPUT = '[data-test-id="formfield-firstName-input"]';
-exports.FIRST_NAME_ERROR_MESSAGE = '[data-test-id="firstnameErrorMessage"]';
+const FIRST_NAME_INPUT = '[data-test-id="formfield-firstName-input"]';
+const FIRST_NAME_ERROR_MESSAGE = '[data-test-id="firstnameErrorMessage"]';
 
-exports.LAST_NAME_INPUT = '[data-test-id="formfield-lastName-input"]';
-exports.LAST_NAME_ERROR_MESSAGE = '[data-test-id="lastnameErrorMessage"]';
+const LAST_NAME_INPUT = '[data-test-id="formfield-lastName-input"]';
+const LAST_NAME_ERROR_MESSAGE = '[data-test-id="lastnameErrorMessage"]';
 
-exports.EMAIL_INPUT = '[data-test-id="formfield-email-input"]';
-exports.EMAIL_ERROR_MESSAGE = '[data-test-id="emailErrorMessage"]';
+const EMAIL_INPUT = '[data-test-id="formfield-email-input"]';
+const EMAIL_ERROR_MESSAGE = '[data-test-id="emailErrorMessage"]';
 
-exports.PASSWORD_INPUT = '[data-test-id="formfield-password-input"]';
-exports.PASSWORD_ERROR_MESSAGE = '[data-test-id="passwordErrorMessage"]';
+const PASSWORD_INPUT = '[data-test-id="formfield-password-input"]';
+const PASSWORD_ERROR_MESSAGE = '[data-test-id="passwordErrorMessage"]';
+
+export {
+    CREATE_ACCOUNT_BUTTON,
+    TERMS_AND_CONDITIONS_LINK,
+    PRIVACY_POLICY_LINK,
+    COOKIES_POLICY_LINK,
+    REQUEST_ERROR,
+    FIRST_NAME_INPUT,
+    FIRST_NAME_ERROR_MESSAGE,
+    LAST_NAME_INPUT,
+    LAST_NAME_ERROR_MESSAGE,
+    EMAIL_INPUT,
+    EMAIL_ERROR_MESSAGE,
+    PASSWORD_INPUT,
+    PASSWORD_ERROR_MESSAGE
+};