+++ packages/components/pages/f-takeawaypay-activation/stories/TakeawaypayActivation.stories.js
@@ -1,14 +1,12 @@
-import {
-    withKnobs, text, select
-} from '@storybook/addon-knobs';
 import { withA11y } from '@storybook/addon-a11y';
 import { locales } from '@justeat/storybook/constants/globalisation';
+import { AUTHENTICATION_JWT } from '../test-utils/constants/f-takeawaypayActivation';
 import TakeawaypayActivation from '../src/components/TakeawaypayActivation.vue';
 import ApiMock from '../src/demo/apiMock';
 
 export default {
     title: 'Components/Pages',
-    decorators: [withKnobs, withA11y]
+    decorators: [withA11y]
 };
 
 const activationStatusAvailableResponse = '/activation-status-200.json';
@@ -32,32 +30,6 @@ const activationBadRequest = '400';
 const activationInternalError = '500';
 const activateSuccessDescription = 'Activation successful';
 
-const authenticationLoggedInRegistered = 'Logged In (registered)';
-const authenticationLoggedInGuest = 'Logged In (guest)';
-
-// eslint-disable-next-line
-const mockAuthTokenRegistered = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.'
-  + 'eyJlbWFpbCI6ImpvZS5ibG9nZ3NAanVzdGVhdHRha2Vhd2F5LmNvbS'
-  + 'IsImNyZWF0ZWRfZGF0ZSI6IjIwMjEtMDItMDhUMTA6Mjc6NDkuMTkz'
-  + 'MDAwMFoiLCJuYW1lIjoiSm9lIEJsb2dncyIsImdsb2JhbF91c2VyX2lkI'
-  + 'joiVTdOUkFsV0FnNXpPZHNkUmdmN25rVHlvaTkwWEVvPSIsImdpdmVuX25h'
-  + 'bWUiOiJKb2UiLCJmYW1pbHlfbmFtZSI6IkJsb2dncyIsImlhdCI6MTYxNTQ2OTUxNn0.VapH6uHnn4lHIkvN_mS9A9IVVWL0YPNE39gDDD-l7SU';
-
-// eslint-disable-next-line
-const mockAuthTokenGuest = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.'
-  + 'eyJlbWFpbCI6ImpvZS5ibG9nZ3NAanVzdGVhdHRha2Vhd2F5LmNvbSIsImN'
-  + 'yZWF0ZWRfZGF0ZSI6IjIwMjEtMDItMDhUMTA6Mjc6NDkuMTkzMDAwMFoiLCJ'
-  + 'zdWIiOiIxODcwMzA5MyIsIm5hbWUiOiJKb2UgQmxvZ2dzIiwiZ2xvYmFsX3V'
-  + 'zZXJfaWQiOiJVN05SQWxXQWc1ek9kc2RSZ2Y3bmtUeW9pOTBYRW89IiwiZ2l2'
-  + 'ZW5fbmFtZSI6IkpvZSIsImZhbWlseV9uYW1lIjoiQmxvZ2dzIiwicm9sZSI6Ik'
-  + 'd1ZXN0IiwiaWF0IjoxNjE1NDY5NTE2fQ.d-dLpZM8vi7IR1GdqZI9IPzUcjnidU-7qO62B3Nfk6I';
-
-const authenticationOptions = {
-    None: null,
-    [authenticationLoggedInRegistered]: mockAuthTokenRegistered,
-    [authenticationLoggedInGuest]: mockAuthTokenGuest
-};
-
 const activationStatusOptions = {
     None: null,
     [activationBadRequestDescription]: activationBadRequest,
@@ -70,37 +42,12 @@ const activateOptions = {
     [activationInternalErrorDescription]: activationInternalError
 };
 
-export const TakeawaypayActivationComponent = () => ({
+export const TakeawaypayActivationComponent = (args, { argTypes }) => ({
     components: { TakeawaypayActivation },
-    props: {
-        locale: {
-            default: select('Locale', [locales.gb, locales.au, locales.nz, locales.ie, locales.es, locales.it], locales.gb)
-        },
-        authentication: {
-            default: select('Authentication', authenticationOptions)
-        },
-        activationStatusResponse: {
-            default: select('Activation Status Response', activationStatusOptions)
-        },
-        activateResponse: {
-            default: select('Activate Takeaway Pay Response', activateOptions)
-        },
-        loginUrl: {
-            default: text('Login URL', '/account/login')
-        },
-        registrationUrl: {
-            default: text('Registration URL', '/account/register')
-        },
-        homeUrl: {
-            default: text('Home URL', '/home')
-        },
-        employeeId: {
-            default: text('Employee Id', '12345')
-        }
-    },
+    props: Object.keys(argTypes),
     computed: {
         authToken () {
-            return this.authentication ? this.authentication : '';
+            return this.isLoggedIn ? AUTHENTICATION_JWT : '';
         },
         getActivationStatusUrl () {
             return this.activationStatusResponse ? `/activation-status-${this.activationStatusResponse}.json` : '/activation-status-200.json';
@@ -123,5 +70,54 @@ export const TakeawaypayActivationComponent = () => ({
             // eslint-disable-next-line no-template-curly-in-string
             ':key="`${authToken},${getActivationStatusUrl},${activationUrl},${locale}`" />'
 });
+
+TakeawaypayActivationComponent.argTypes = {
+    locale: {
+        control: { type: 'select' },
+        options: [locales.gb, locales.au, locales.nz, locales.it, locales.es, locales.ie],
+        description: 'Locale'
+    },
+    isLoggedIn: {
+        control: { type: 'boolean' },
+        description: 'Is User Logged In'
+    },
+    activationStatusResponse: {
+        control: { type: 'select' },
+        options: activationStatusOptions,
+        description: 'Activation Status Response'
+    },
+    activateResponse: {
+        control: { type: 'select' },
+        options: activateOptions,
+        description: 'Activate Takeaway Pay Response'
+    },
+    loginUrl: {
+        control: { type: 'text' },
+        description: 'Login URL'
+    },
+    registrationUrl: {
+        control: { type: 'text' },
+        description: 'Registration URL'
+    },
+    homeUrl: {
+        control: { type: 'text' },
+        description: 'Home URL'
+    },
+    employeeId: {
+        control: { type: 'text' },
+        description: 'Employee ID'
+    }
+};
+
+TakeawaypayActivationComponent.args = {
+    employeeId: '12345',
+    homeUrl: '/home',
+    registrationUrl: '/account/register',
+    loginUrl: '/account/login',
+    activateResponse: null,
+    activationStatusResponse: null,
+    locale: locales.gb,
+    isLoggedIn: false
+};
 
 TakeawaypayActivationComponent.storyName = 'f-takeawaypay-activation';