+++ packages/components/pages/f-takeawaypay-activation/test-utils/component-objects/f-takeawaypayActivation-error.component.js
@@ -1,0 +1,12 @@
+import Page from '@justeat/f-wdio-utils';
+import { ERROR_COMPONENT } from './f-takeawaypayActivation-selectors';
+
+class TakeawaypayActivationError extends Page {
+    constructor () {
+        super('page', 'takeawaypay-activation-component');
+    }
+
+    get component () { return $(ERROR_COMPONENT); }
+}
+
+export default new TakeawaypayActivationError();